moved torch_test to gpu_test, added test for TF support
